requirements.txt
PyQt5>=5.15.0
pyyaml>=5.1
requests>=2.25.0
cryptography>=3.4.0